all the refactoring
