+++ frontend/src/Metamaps/Realtime/events.js
@@ -1,0 +1,56 @@
+/* EVENTS SENDABLE */
+export const REQUEST_LIVE_MAPS = 'REQUEST_LIVE_MAPS'
+export const JOIN_MAP = 'JOIN_MAP'
+export const LEAVE_MAP = 'LEAVE_MAP'
+export const CHECK_FOR_CALL = 'CHECK_FOR_CALL'
+export const ACCEPT_CALL = 'ACCEPT_CALL' 
+export const DENY_CALL = 'DENY_CALL'
+export const DENY_INVITE = 'DENY_INVITE'
+export const INVITE_TO_JOIN = 'INVITE_TO_JOIN'
+export const INVITE_A_CALL = 'INVITE_A_CALL'
+export const JOIN_CALL = 'JOIN_CALL'
+export const LEAVE_CALL = 'LEAVE_CALL'
+export const REQUEST_MAPPER_INFO = 'REQUEST_MAPPER_INFO' 
+export const SEND_MAPPER_INFO = 'SEND_MAPPER_INFO'
+export const SEND_COORDS = 'SEND_COORDS' 
+export const CREATE_MESSAGE = 'CREATE_MESSAGE' 
+export const DRAG_TOPIC = 'DRAG_TOPIC'
+export const CREATE_TOPIC = 'CREATE_TOPIC'
+export const UPDATE_TOPIC = 'UPDATE_TOPIC'
+export const REMOVE_TOPIC = 'REMOVE_TOPIC'
+export const DELETE_TOPIC = 'DELETE_TOPIC'
+export const CREATE_SYNAPSE = 'CREATE_SYNAPSE'
+export const UPDATE_SYNAPSE = 'UPDATE_SYNAPSE'
+export const REMOVE_SYNAPSE = 'REMOVE_SYNAPSE'
+export const DELETE_SYNAPSE = 'DELETE_SYNAPSE'
+export const UPDATE_MAP = 'UPDATE_MAP' 
+
+/* EVENTS RECEIVABLE */
+export const INVITED_TO_CALL = 'INVITED_TO_CALL'
+export const INVITED_TO_JOIN = 'INVITED_TO_JOIN'
+export const CALL_ACCEPTED = 'CALL_ACCEPTED'
+export const CALL_DENIED = 'CALL_DENIED'
+export const INVITE_DENIED = 'INVITE_DENIED'
+export const CALL_IN_PROGRESS = 'CALL_IN_PROGRESS'
+export const CALL_STARTED = 'CALL_STARTED'
+export const MAPPER_JOINED_CALL = 'MAPPER_JOINED_CALL'
+export const MAPPER_LEFT_CALL = 'MAPPER_LEFT_CALL'
+export const MAPPER_LIST_UPDATED = 'MAPPER_LIST_UPDATED'
+export const NEW_MAPPER = 'NEW_MAPPER'
+export const LOST_MAPPER = 'LOST_MAPPER'
+export const MESSAGE_CREATED = 'MESSAGE_CREATED'
+export const TOPIC_DRAGGED = 'TOPIC_DRAGGED'
+export const TOPIC_CREATED = 'TOPIC_CREATED'
+export const TOPIC_UPDATED = 'TOPIC_UPDATED'
+export const TOPIC_REMOVED = 'TOPIC_REMOVED'
+export const TOPIC_DELETED = 'TOPIC_DELETED'
+export const SYNAPSE_CREATED = 'SYNAPSE_CREATED'
+export const SYNAPSE_UPDATED = 'SYNAPSE_UPDATED'
+export const SYNAPSE_REMOVED = 'SYNAPSE_REMOVED'
+export const SYNAPSE_DELETED = 'SYNAPSE_DELETED'
+export const PEER_COORDS_UPDATED = 'PEER_COORDS_UPDATED'
+export const MAP_UPDATED = 'MAP_UPDATED'
+export const LIVE_MAPS_RECEIVED = 'LIVE_MAPS_RECEIVED'
+export const MAP_WENT_LIVE = 'MAP_WENT_LIVE'
+export const MAP_CEASED_LIVE = 'MAP_CEASED_LIVE'
+
